requirements.txt
asgiref==3.8.1
coverage==7.5.2
Django==5.0.6
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-magic==0.4.27
sqlparse==0.5.0
tzdata==2024.1
